refactor: Simplify recipe registration by consolidating waystone logic into RecipeManager and removing WaystoneRecipe class
